fix: fix a error caused by returning a NoneType
There's a chance that this query returns a NoneType because there's no
guarantee that an object with the provided id actually exists in the
database. So, when runing the .one() function, server will throw an
error page if there isn’t an object with this id.

To mitigate this issue, use the one_or_none() function. Instead of
throwing an error, this function will simply return a NoneType object.
Then, using a logic to see if the query returned a NoneType, if it did,
redirect to "No such item!!" page.